+++ CHANGELOG.md
@@ -19,6 +19,8 @@ owns this process itself by writing `roff` directly.
 
 Performance improvements:
 
+* [PERF #1746](https://github.com/BurntSushi/ripgrep/issues/1746):
+  Make some cases with inner literals faster.
 * [PERF #1760](https://github.com/BurntSushi/ripgrep/issues/1760):
   Make most searches with `\b` look-arounds (among others) much faster.
 * [PERF #2591](https://github.com/BurntSushi/ripgrep/pull/2591):
@@ -30,16 +32,30 @@ Feature enhancements:
 
 * Added or improved file type filtering for Ada, DITA, Elixir, Fuchsia, Gentoo,
   Gradle, GraphQL, Markdown, Prolog, Raku, TypeScript, USD, V
+* [FEATURE #1709](https://github.com/BurntSushi/ripgrep/issues/1709):
+  Improve documentation of ripgrep's behavior when stdout is a tty.
+* [FEATURE #1737](https://github.com/BurntSushi/ripgrep/issues/1737):
+  Provide binaries for Apple silicon.
 * [FEATURE #1790](https://github.com/BurntSushi/ripgrep/issues/1790):
   Add new `--stop-on-nonmatch` flag.
 * [FEATURE #1814](https://github.com/BurntSushi/ripgrep/issues/1814):
   Flags are now categorized in `-h/--help` output and ripgrep's man page.
+* [FEATURE #1838](https://github.com/BurntSushi/ripgrep/issues/1838):
+  An error is shown when searching for NUL bytes with binary detection enabled.
 * [FEATURE #2195](https://github.com/BurntSushi/ripgrep/issues/2195):
   When `extra-verbose` mode is enabled in zsh, show extra file type info.
+* [FEATURE #2298](https://github.com/BurntSushi/ripgrep/issues/2298):
+  Add instructions for installing ripgrep using `cargo binstall`.
 * [FEATURE #2409](https://github.com/BurntSushi/ripgrep/pull/2409):
   Added installation instructions for `winget`.
+* [FEATURE #2425](https://github.com/BurntSushi/ripgrep/pull/2425):
+  Shell completions (and man page) can be created via `rg --generate`.
+* [FEATURE #2524](https://github.com/BurntSushi/ripgrep/issues/2524):
+  The `--debug` flag now indicates whether stdin or `./` is being searched.
 * [FEATURE #2643](https://github.com/BurntSushi/ripgrep/issues/2643):
   Make `-d` a short flag for `--max-depth`.
+* [FEATURE #2645](https://github.com/BurntSushi/ripgrep/issues/2645):
+  The `--version` output will now also contain PCRE2 availability information.
 
 Bug fixes:
 
@@ -47,6 +63,10 @@ Bug fixes:
   Don't error when `-v/--invert-match` is used multiple times.
 * [BUG #1275](https://github.com/BurntSushi/ripgrep/issues/1275):
   Fix bug with `\b` assertion in the regex engine.
+* [BUG #1376](https://github.com/BurntSushi/ripgrep/issues/1376):
+  Using `--no-ignore --ignore-vcs` now works as one would expect.
+* [BUG #1622](https://github.com/BurntSushi/ripgrep/issues/1622):
+  Add note about error messages to `-z/--search-zip` documentation.
 * [BUG #1648](https://github.com/BurntSushi/ripgrep/issues/1648):
   Fix bug where sometimes short flags with values, e.g., `-M 900`, would fail.
 * [BUG #1701](https://github.com/BurntSushi/ripgrep/issues/1701):
@@ -59,22 +79,40 @@ Bug fixes:
   Disable mmap searching in all non-64-bit environments.
 * [BUG #1966](https://github.com/BurntSushi/ripgrep/issues/1966):
   Fix bug where ripgrep can panic when printing to stderr.
+* [BUG #2046](https://github.com/BurntSushi/ripgrep/issues/2046):
+  Clarify that `--pre` can accept any kind of path in the documentation.
 * [BUG #2108](https://github.com/BurntSushi/ripgrep/issues/2108):
   Improve docs for `-r/--replace` syntax.
 * [BUG #2198](https://github.com/BurntSushi/ripgrep/issues/2198):
   Fix bug where `--no-ignore-dot` would not ignore `.rgignore`.
+* [BUG #2201](https://github.com/BurntSushi/ripgrep/issues/2201):
+  Improve docs for `-r/--replace` flag.
 * [BUG #2288](https://github.com/BurntSushi/ripgrep/issues/2288):
   `-A` and `-B` now only each partially override `-C`.
 * [BUG #2236](https://github.com/BurntSushi/ripgrep/issues/2236):
   Fix gitignore parsing bug where a trailing `\/` resulted in an error.
 * [BUG #2243](https://github.com/BurntSushi/ripgrep/issues/2243):
   Fix `--sort` flag for values other than `path`.
+* [BUG #2246](https://github.com/BurntSushi/ripgrep/issues/2246):
+  Add note in `--debug` logs when binary files are ignored.
+* [BUG #2337](https://github.com/BurntSushi/ripgrep/issues/2337):
+  Improve docs to mention that `--stats` is always implied by `--json`.
 * [BUG #2381](https://github.com/BurntSushi/ripgrep/issues/2381):
   Make `-p/--pretty` override flags like `--no-line-number`.
 * [BUG #2392](https://github.com/BurntSushi/ripgrep/issues/2392):
   Improve global git config parsing of the `excludesFile` field.
+* [BUG #2418](https://github.com/BurntSushi/ripgrep/pull/2418):
+  Clarify sorting semantics of `--sort=path`.
+* [BUG #2458](https://github.com/BurntSushi/ripgrep/pull/2458):
+  Make `--trim` run before `-M/--max-columns` takes effect.
+* [BUG #2479](https://github.com/BurntSushi/ripgrep/issues/2479):
+  Add documentation about `.ignore`/`.rgignore` files in parent directories.
 * [BUG #2480](https://github.com/BurntSushi/ripgrep/issues/2480):
   Fix bug when using inline regex flags with `-e/--regexp`.
+* [BUG #2505](https://github.com/BurntSushi/ripgrep/issues/2505):
+  Improve docs for `--vimgrep` by mentioning footguns and some work-arounds.
+* [BUG #2519](https://github.com/BurntSushi/ripgrep/issues/2519):
+  Fix incorrect default value in documentation for `--field-match-separator`.
 * [BUG #2523](https://github.com/BurntSushi/ripgrep/issues/2523):
   Make executable searching take `.com` into account on Windows.
 * [BUG #2574](https://github.com/BurntSushi/ripgrep/issues/2574):